add jwt tokenization

+++ resources/item.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required,get_jwt
 from sqlalchemy.exc import SQLAlchemyError   
 
 from db import db
@@ -15,13 +14,19 @@ blp = Blueprint("Items",__name__,description="Operations on items")
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
     # Retrieve an item by its ID
+    @jwt_required()
     @blp.response(200,ItemSchema)
     def get(self,item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
     # Delete an item by its ID
+    @jwt_required()
     def delete(self,item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401,message="Admin privilege required.")
+            
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -49,11 +54,13 @@ class Item(MethodView):
 @blp.route("/item")
 class ItemList(MethodView):
     # Retrieve all items
+    @jwt_required()
     @blp.response(200,ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
     # Create a new item
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
